feat: complete backend entities according to DER
Core Entities:
- Orders & OrderItems (with price history tracking)
- Cart & CartItems (with audit timestamps)
- Payments (with method/status enums)
- Offers & OfferProducts (with discounted prices)
- Notifications (with type classification)
- Users, Roles, Addresses, Products, Variants

Technical Improvements:
- BigDecimal precision=19, scale=2 for monetary fields
- FetchType.LAZY for all @ManyToOne relationships
- Proper package structure and imports
- UUID externalId for all entities
- CreatedAt timestamps for business-relevant entities

Enums Created:
- PaymentMethod (5 payment types)
- PaymentStatus (4 payment states)
- NotificationType (8 notification categories)

+++ src/main/java/com/utn/corralon/features/orders/orderEntity.java
@@ -1,10 +1,11 @@
 package com.utn.corralon.features.orders;
 
 import com.utn.corralon.features.addresses.addressEntity;
-import com.utn.corralon.features.user.userEntity;
+import com.utn.corralon.features.users.userEntity;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -34,7 +35,7 @@ public class orderEntity {
     private addressEntity address;
 
     @Column(name="total",nullable = false, precision = 19, scale = 2)
-    private Double total;
+    private BigDecimal total;
 
     @Column(name="created_at",nullable = false)
     private LocalDateTime created_at;
